requirements.txt
streamlit==1.16.0
numpy
matplotlib
scipy
Pillow
streamlit-extras==0.2.4
PyQt5==5.15.4